requirements.txt
# Core Web Framework
fastapi==0.124.4
uvicorn==0.33.0
pydantic==2.10.6
python-dotenv==1.0.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.15
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0

# RAG & Embeddings
faiss-cpu>=1.9.0
sentence-transformers>=2.6.0
